Shifted StoryActivity tests
