+++ app/src/sharedTest/java/org/oppia/android/app/story/StoryFragmentTest.kt
@@ -33,6 +33,7 @@ import dagger.Component
 import dagger.Module
 import dagger.Provides
 import org.hamcrest.CoreMatchers.allOf
+import org.hamcrest.CoreMatchers.not
 import org.hamcrest.Description
 import org.hamcrest.Matcher
 import org.hamcrest.TypeSafeMatcher
@@ -65,6 +66,7 @@ import org.oppia.android.app.recyclerview.RecyclerViewMatcher.Companion.atPositi
 import org.oppia.android.app.recyclerview.RecyclerViewMatcher.Companion.atPositionOnView
 import org.oppia.android.app.recyclerview.RecyclerViewMatcher.Companion.hasItemCount
 import org.oppia.android.app.shim.ViewBindingShimModule
+import org.oppia.android.app.utility.EspressoTestsMatchers.withDrawable
 import org.oppia.android.app.utility.OrientationChangeAction.Companion.orientationLandscape
 import org.oppia.android.app.utility.anyOrNull
 import org.oppia.android.app.utility.capture
@@ -87,6 +89,8 @@ import org.oppia.android.domain.question.QuestionModule
 import org.oppia.android.domain.topic.FRACTIONS_STORY_ID_0
 import org.oppia.android.domain.topic.FRACTIONS_TOPIC_ID
 import org.oppia.android.domain.topic.PrimeTopicAssetsControllerModule
+import org.oppia.android.domain.topic.RATIOS_STORY_ID_0
+import org.oppia.android.domain.topic.RATIOS_TOPIC_ID
 import org.oppia.android.domain.topic.TEST_STORY_ID_1
 import org.oppia.android.domain.topic.TEST_TOPIC_ID_1
 import org.oppia.android.testing.RobolectricModule
@@ -537,6 +541,112 @@ class StoryFragmentTest {
     }
   }
 
+  @Config(qualifiers = "+sw600dp")
+  @Test
+  fun testStoryFragment_completedChapter_checkProgressDrawableIsCorrect() {
+    setStoryPartialProgressForFractions()
+    launch<StoryActivity>(createFractionsStoryActivityIntent()).use {
+      testCoroutineDispatchers.runCurrent()
+      onView(
+        atPositionOnView(
+          recyclerViewId = R.id.story_chapter_list,
+          position = 1,
+          targetViewId = R.id.progress_image_view
+        )
+      ).check(matches(withDrawable(R.drawable.circular_solid_color_primary_32dp)))
+    }
+  }
+
+  @Config(qualifiers = "+sw600dp")
+  @Test
+  fun testStoryFragment_notStartedChapter_checkProgressDrawableIsCorrect() {
+    launch<StoryActivity>(createFractionsStoryActivityIntent()).use {
+      testCoroutineDispatchers.runCurrent()
+      onView(
+        atPositionOnView(
+          recyclerViewId = R.id.story_chapter_list,
+          position = 1,
+          targetViewId = R.id.progress_image_view
+        )
+      ).check(matches(withDrawable(R.drawable.circular_stroke_2dp_color_primary_32dp)))
+    }
+  }
+
+  @Config(qualifiers = "+sw600dp")
+  @Test
+  fun testStoryFragment_lockedChapter_checkProgressDrawableIsCorrect() {
+    launch<StoryActivity>(createRatiosStoryActivityIntent()).use {
+      testCoroutineDispatchers.runCurrent()
+      onView(
+        atPositionOnView(
+          recyclerViewId = R.id.story_chapter_list,
+          position = 1,
+          targetViewId = R.id.progress_image_view
+        )
+      ).check(matches(withDrawable(R.drawable.circular_stroke_1dp_grey_32dp)))
+    }
+  }
+
+  @Config(qualifiers = "+sw600dp")
+  @Test
+  fun testStoryFragment_completedChapter_pawIconIsVisible() {
+    launch<StoryActivity>(createFractionsStoryActivityIntent()).use {
+      testCoroutineDispatchers.runCurrent()
+      onView(
+        atPositionOnView(
+          recyclerViewId = R.id.story_chapter_list,
+          position = 1,
+          targetViewId = R.id.completed_chapter_image_view
+        )
+      ).check(matches(withDrawable(R.drawable.ic_lessons_icon_24dp)))
+    }
+  }
+
+  @Config(qualifiers = "+sw600dp")
+  @Test
+  fun testStoryFragment_pendingChapter_pawIconIsGone() {
+    launch<StoryActivity>(createFractionsStoryActivityIntent()).use {
+      testCoroutineDispatchers.runCurrent()
+      onView(
+        atPositionOnView(
+          recyclerViewId = R.id.story_chapter_list,
+          position = 2,
+          targetViewId = R.id.completed_chapter_image_view
+        )
+      ).check(matches(not(isDisplayed())))
+    }
+  }
+
+  @Config(qualifiers = "+sw600dp")
+  @Test
+  fun testStoryFragment_completedChapter_verticalDashedLineIsVisible() {
+    launch<StoryActivity>(createFractionsStoryActivityIntent()).use {
+      testCoroutineDispatchers.runCurrent()
+      onView(
+        atPositionOnView(
+          recyclerViewId = R.id.story_chapter_list,
+          position = 1,
+          targetViewId = R.id.verticalDashedLineView
+        )
+      ).check(matches(isDisplayed()))
+    }
+  }
+
+  @Config(qualifiers = "+sw600dp")
+  @Test
+  fun testStoryFragment_lastChapter_verticalDashedLineIsGone() {
+    launch<StoryActivity>(createFractionsStoryActivityIntent()).use {
+      testCoroutineDispatchers.runCurrent()
+      onView(
+        atPositionOnView(
+          recyclerViewId = R.id.story_chapter_list,
+          position = 2,
+          targetViewId = R.id.verticalDashedLineView
+        )
+      ).check(matches(not(isDisplayed())))
+    }
+  }
+
   /**
    * Returns an action that finds a TextView containing the specific text, finds a ClickableSpan
    * within that text view that contains the specified text, then clicks it. The need for this was
@@ -602,6 +712,15 @@ class StoryFragmentTest {
       internalProfileId,
       TEST_TOPIC_ID_1,
       TEST_STORY_ID_1
+    )
+  }
+
+  private fun createRatiosStoryActivityIntent(): Intent {
+    return StoryActivity.createStoryActivityIntent(
+      ApplicationProvider.getApplicationContext(),
+      internalProfileId,
+      RATIOS_TOPIC_ID,
+      RATIOS_STORY_ID_0
     )
   }
