+++ app/src/test/java/org/oppia/android/app/story/StoryActivityLocalTest.kt
@@ -1,24 +1,18 @@
 package org.oppia.android.app.story
 
 import android.app.Application
-import android.content.Context
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
-import androidx.test.core.app.ActivityScenario.launch
+import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
 import dagger.Component
-import org.hamcrest.CoreMatchers.not
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.oppia.android.R
 import org.oppia.android.app.activity.ActivityComponent
 import org.oppia.android.app.application.ActivityComponentFactory
 import org.oppia.android.app.application.ApplicationComponent
@@ -28,11 +22,8 @@ import org.oppia.android.app.application.ApplicationModule
 import org.oppia.android.app.application.ApplicationStartupListenerModule
 import org.oppia.android.app.model.EventLog
 import org.oppia.android.app.model.EventLog.Context.ActivityContextCase.STORY_CONTEXT
-import org.oppia.android.app.model.ProfileId
 import org.oppia.android.app.player.state.hintsandsolution.HintsAndSolutionConfigModule
-import org.oppia.android.app.recyclerview.RecyclerViewMatcher.Companion.atPositionOnView
 import org.oppia.android.app.shim.ViewBindingShimModule
-import org.oppia.android.app.utility.EspressoTestsMatchers.withDrawable
 import org.oppia.android.domain.classify.InteractionsModule
 import org.oppia.android.domain.classify.rules.continueinteraction.ContinueModule
 import org.oppia.android.domain.classify.rules.dragAndDropSortInput.DragDropSortInputModule
@@ -49,19 +40,11 @@ import org.oppia.android.domain.oppialogger.LogStorageModule
 import org.oppia.android.domain.oppialogger.loguploader.LogUploadWorkerModule
 import org.oppia.android.domain.oppialogger.loguploader.WorkManagerConfigurationModule
 import org.oppia.android.domain.question.QuestionModule
-import org.oppia.android.domain.topic.FRACTIONS_STORY_ID_0
-import org.oppia.android.domain.topic.FRACTIONS_TOPIC_ID
 import org.oppia.android.domain.topic.PrimeTopicAssetsControllerModule
-import org.oppia.android.domain.topic.RATIOS_STORY_ID_0
-import org.oppia.android.domain.topic.RATIOS_TOPIC_ID
 import org.oppia.android.testing.FakeEventLogger
 import org.oppia.android.testing.RobolectricModule
-import org.oppia.android.testing.TestCoroutineDispatchers
 import org.oppia.android.testing.TestDispatcherModule
 import org.oppia.android.testing.TestLogReportingModule
-import org.oppia.android.testing.profile.ProfileTestHelper
-import org.oppia.android.testing.story.StoryProgressTestHelper
-import org.oppia.android.testing.time.FakeOppiaClock
 import org.oppia.android.testing.time.FakeOppiaClockModule
 import org.oppia.android.util.accessibility.AccessibilityTestModule
 import org.oppia.android.util.caching.testing.CachingTestModule
@@ -83,53 +66,29 @@ private const val TEST_STORY_ID = "GJ2rLXRKD5hw"
 @LooperMode(LooperMode.Mode.PAUSED)
 @Config(
   application = StoryActivityLocalTest.TestApplication::class,
-  qualifiers = "sw600dp"
+  qualifiers = "port-xxhdpi"
 )
 class StoryActivityLocalTest {
 
   @Inject
   lateinit var fakeEventLogger: FakeEventLogger
 
-  @Inject
-  lateinit var profileTestHelper: ProfileTestHelper
-
-  @Inject
-  lateinit var storyProgressTestHelper: StoryProgressTestHelper
-
-  @Inject
-  lateinit var context: Context
-
-  @Inject
-  lateinit var testCoroutineDispatchers: TestCoroutineDispatchers
-
-  @Inject
-  lateinit var fakeOppiaClock: FakeOppiaClock
-
   private val internalProfileId = 0
-  private lateinit var profileId: ProfileId
 
   @Before
   fun setUp() {
     Intents.init()
     setUpTestApplicationComponent()
-    testCoroutineDispatchers.registerIdlingResource()
-    profileId = ProfileId.newBuilder().setInternalId(internalProfileId).build()
-    fakeOppiaClock.setFakeTimeMode(FakeOppiaClock.FakeTimeMode.MODE_UPTIME_MILLIS)
-    storyProgressTestHelper.markCompletedFractionsStory0Exp0(
-      profileId,
-      timestampOlderThanOneWeek = false
-    )
   }
 
   @After
   fun tearDown() {
     Intents.release()
-    testCoroutineDispatchers.unregisterIdlingResource()
   }
 
   @Test
   fun testStoryActivity_onLaunch_logsEvent() {
-    launch<StoryActivity>(
+    ActivityScenario.launch<StoryActivity>(
       createStoryActivityIntent(internalProfileId, TEST_TOPIC_ID, TEST_STORY_ID)
     ).use {
       val event = fakeEventLogger.getMostRecentEvent()
@@ -139,118 +98,6 @@ class StoryActivityLocalTest {
       assertThat(event.context.activityContextCase).isEqualTo(STORY_CONTEXT)
       assertThat(event.context.storyContext.storyId).matches(TEST_STORY_ID)
       assertThat(event.context.storyContext.topicId).matches(TEST_TOPIC_ID)
-    }
-  }
-
-  @Test
-  fun testStoryFragment_completedChapter_checkProgressDrawableIsCorrect() {
-    launch<StoryActivity>(
-      createStoryActivityIntent(internalProfileId, FRACTIONS_TOPIC_ID, FRACTIONS_STORY_ID_0)
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(
-        atPositionOnView(
-          recyclerViewId = R.id.story_chapter_list,
-          position = 1,
-          targetViewId = R.id.progress_image_view
-        )
-      ).check(matches(withDrawable(R.drawable.circular_solid_color_primary_32dp)))
-    }
-  }
-
-  @Test
-  fun testStoryFragment_notStartedChapter_checkProgressDrawableIsCorrect() {
-    launch<StoryActivity>(
-      createStoryActivityIntent(internalProfileId, FRACTIONS_TOPIC_ID, FRACTIONS_STORY_ID_0)
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(
-        atPositionOnView(
-          recyclerViewId = R.id.story_chapter_list,
-          position = 1,
-          targetViewId = R.id.progress_image_view
-        )
-      ).check(matches(withDrawable(R.drawable.circular_stroke_2dp_color_primary_32dp)))
-    }
-  }
-
-  @Test
-  fun testStoryFragment_lockedChapter_checkProgressDrawableIsCorrect() {
-    launch<StoryActivity>(
-      createStoryActivityIntent(internalProfileId, RATIOS_TOPIC_ID, RATIOS_STORY_ID_0)
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(
-        atPositionOnView(
-          recyclerViewId = R.id.story_chapter_list,
-          position = 1,
-          targetViewId = R.id.progress_image_view
-        )
-      ).check(matches(withDrawable(R.drawable.circular_stroke_1dp_grey_32dp)))
-    }
-  }
-
-  @Test
-  fun testStoryFragment_completedChapter_pawIconIsVisible() {
-    launch<StoryActivity>(
-      createStoryActivityIntent(internalProfileId, FRACTIONS_TOPIC_ID, FRACTIONS_STORY_ID_0)
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(
-        atPositionOnView(
-          recyclerViewId = R.id.story_chapter_list,
-          position = 1,
-          targetViewId = R.id.completed_chapter_image_view
-        )
-      ).check(matches(withDrawable(R.drawable.ic_lessons_icon_24dp)))
-    }
-  }
-
-  @Test
-  fun testStoryFragment_pendingChapter_pawIconIsGone() {
-    launch<StoryActivity>(
-      createStoryActivityIntent(internalProfileId, FRACTIONS_TOPIC_ID, FRACTIONS_STORY_ID_0)
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(
-        atPositionOnView(
-          recyclerViewId = R.id.story_chapter_list,
-          position = 2,
-          targetViewId = R.id.completed_chapter_image_view
-        )
-      ).check(matches(not(isDisplayed())))
-    }
-  }
-
-  @Test
-  fun testStoryFragment_completedChapter_verticalDashedLineIsVisible() {
-    launch<StoryActivity>(
-      createStoryActivityIntent(internalProfileId, FRACTIONS_TOPIC_ID, FRACTIONS_STORY_ID_0)
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(
-        atPositionOnView(
-          recyclerViewId = R.id.story_chapter_list,
-          position = 1,
-          targetViewId = R.id.verticalDashedLineView
-        )
-      ).check(matches(isDisplayed()))
-    }
-  }
-
-  @Test
-  fun testStoryFragment_lastChapter_verticalDashedLineIsGone() {
-    launch<StoryActivity>(
-      createStoryActivityIntent(internalProfileId, FRACTIONS_TOPIC_ID, FRACTIONS_STORY_ID_0)
-    ).use {
-      testCoroutineDispatchers.runCurrent()
-      onView(
-        atPositionOnView(
-          recyclerViewId = R.id.story_chapter_list,
-          position = 2,
-          targetViewId = R.id.verticalDashedLineView
-        )
-      ).check(matches(not(isDisplayed())))
     }
   }
